requirements.txt
gradio
openai
python-dotenv
pypdf
python-docx
pdfkit
requests
flask